# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio-Fixer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/posts/{slug}:
    get:
      operationId: getPublicPost
      tags: [blog]
      summary: Get a published blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { $ref: "#/components/schemas/BlogPost" }
                  - $ref: "#/components/schemas/ApiError"

  # ─────────────────────────────────────────────────────────────────────
  # Public contact form
  # ─────────────────────────────────────────────────────────────────────
...
  /v1/admin/posts:
    get:
      operationId: listPosts
      tags: [admin.posts]
      summary: List all posts
      security: [{ adminAuth: [] }]
      parameters:
        - name: userId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { $ref: "#/components/schemas/PaginatedPosts" }
                  - $ref: "#/components/schemas/ApiError"
    post:
      operationId: createPost
      tags: [admin.posts]
      summary: Create a post
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"
  /v1/admin/posts/{id}:
    put:
      operationId: updatePost
      tags: [admin.posts]
      summary: Update a post
      security: [{ adminAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePost
      tags: [admin.posts]
      summary: Delete a post
      security: [{ adminAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"

  # ─────────────────────────────────────────────────────────────────────
  # Admin: messages
  # ─────────────────────────────────────────────────────────────────────
...
  /v1/admin/section-settings:
    get:
      operationId: listSectionSettings
      tags: [admin.section]
      summary: List section settings
      security: [{ adminAuth: [] }]
      responses:
        "200":
          description: Section settings
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data:
                            type: array
                            items: { $ref: "#/components/schemas/SectionSetting" }
                  - $ref: "#/components/schemas/ApiError"
  /v1/admin/section-settings/reorder:
    post:
      operationId: reorderSectionSettings
      tags: [admin.section]
      summary: Reorder section settings
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReorderInput" }
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"
  /v1/admin/section-settings/{id}:
    put:
      operationId: updateSectionSetting
      tags: [admin.section]
      summary: Update a section setting
      security: [{ adminAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SectionSettingInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { type: object, nullable: true }
                  - $ref: "#/components/schemas/ApiError"

  # ─────────────────────────────────────────────────────────────────────
  # Admin: site settings (singleton)
  # ─────────────────────────────────────────────────────────────────────
...
  /v1/admin/cv/settings:
    get:
      operationId: getCvSettings
      tags: [cv]
      summary: Get current CV settings (admin only)
      security: [{ adminAuth: [] }]
      responses:
        "200":
          description: CV settings
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data: { $ref: "#/components/schemas/CvSettings" }
                  - $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateCvSettings
      tags: [cv]
      summary: Save the object path of the uploaded CV
      security: [{ adminAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CvSettingsInput" }
      responses:
        "200":
          description: Settings saved
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data:
                            type: object
                            properties:
                              id: { type: string }
                  - $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteCvSettings
      tags: [cv]
      summary: Delete CV settings and uploaded file
      security: [{ adminAuth: [] }]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/SuccessEnvelope"
                      - type: object
                        properties:
                          data:
                            type: object
                            properties:
                              message: { type: string }
                  - $ref: "#/components/schemas/ApiError"

  # ─────────────────────────────────────────────────────────────────────
  # Admin: preview
  # ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
    ApiError:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, enum: [false] }
        message: { type: string }
        code: { type: string }
        existingId: { type: string }
...
    Pagination:
      type: object
      required: [total, limit, offset, hasMore]
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
...
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, reading_minutes, cover_image_url, tags, is_published, published_at, user_id, deleted_at, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string, nullable: true }
        content: { type: string }
        reading_minutes:
          type: integer
          description: >-
            Stored generated column (migration 063): max(1, ceil(words/200)).
            Read-only - ignored on insert/update.
        cover_image_url: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        is_published: { type: boolean, nullable: true }
        published_at: { type: string, nullable: true }
        user_id: { type: string, nullable: true }
        deleted_at: { type: string, nullable: true }
        created_at: { type: string }
        updated_at: { type: string }
...
    SectionSetting:
      type: object
      required: [id, key, label, is_visible, sort_order, updated_at, created_at]
      properties:
        id: { type: string }
        key: { type: string }
        label: { type: string }
        is_visible: { type: boolean }
        sort_order: { type: integer }
        updated_at: { type: string }
        created_at: { type: string }
...
    CvSettings:
      type: object
      required: [objectPath, fileName, updatedAt]
      properties:
        objectPath: { type: string, nullable: true }
        fileName: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
...
    PaginatedPosts:
      type: object
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/BlogPost" } }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    PostInput:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string, nullable: true }
        content: { type: string }
        cover_image_url: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        is_published: { type: boolean }
...
    SectionSettingInput:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        is_visible: { type: boolean }
        sort_order: { type: integer }
    ReorderInput:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              sort_order: { type: integer }
...
    CvSettingsInput:
      type: object
      required: [objectPath, fileName]
      properties:
        objectPath: { type: string, pattern: '^cv-\d+\.pdf$' }
        fileName: { type: string, minLength: 1, maxLength: 255, pattern: '\.pdf$' }
  securitySchemes:
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin API key for server-side operations. Alternatively, use Bearer token from Clerk JWT.